Working commit ============== + particulary add serialization into db background
git-svn-id: https://wxdbexplorer.svn.sourceforge.net/svnroot/wxdbexplorer/trunk@21 ab01d90b-4ff7-4e84-82d5-2368d4a8d7fa

+++ src/IDbType.h
@@ -10,8 +10,8 @@ public:
 		
 		// size
 		virtual bool HaveSize() = 0;
-		virtual int GetSize() = 0;
-		virtual void SetSize(int size) = 0;
+		virtual long GetSize() = 0;
+		virtual void SetSize(long size) = 0;
 		
 		// primary key
 		virtual bool HavePrimaryKey() = 0;
